Better error handling
Updated UI helper class
Dynamic data in dashboard
Login back button program to navtigate to wallet dashboard after logout
No registration with same exg wallet address

+++ lib/models/campaign/refferals.dart
@@ -1,6 +1,6 @@
 class CampaignRefferals {
   String id;
-  double refferals;
+  int refferals;
   double name;
   double paid;
